+++ pandas/tests/extension/test_boolean.py
@@ -258,7 +258,7 @@ class TestGroupby(base.BaseGroupbyTests):
         _, index = pd.factorize(data_for_grouping, sort=True)
 
         index = pd.Index(index, name="B")
-        expected = pd.Series([3, 1], index=index, name="A")
+        expected = pd.Series([3, 1], dtype="float64", index=index, name="A")
         if as_index:
             self.assert_series_equal(result, expected)
         else:
@@ -271,7 +271,7 @@ class TestGroupby(base.BaseGroupbyTests):
         _, index = pd.factorize(data_for_grouping, sort=False)
 
         index = pd.Index(index, name="B")
-        expected = pd.Series([1, 3], index=index, name="A")
+        expected = pd.Series([1, 3], dtype="float64", index=index, name="A")
         self.assert_series_equal(result, expected)
 
     def test_groupby_extension_transform(self, data_for_grouping):
